pyproject.toml
[tool.poetry]
name = "interview"
version = "0.1.0"
description = ""
authors = ["MO.Norouzi <mo.norouzi@emofid.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
tortoise-orm = "^0.20.0"
APScheduler = "^3.10.4"
uvicorn = "^0.23.2"
python-jose = "^3.3.0"
python-decouple = "^3.8"
python-multipart = "^0.0.6"
pytest = "^7.4.2"
asgi-lifespan = "^2.1.0"
httpx = "^0.25.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
